requirements.txt
dependency-injector==4.39.1
fastapi==0.79.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
sqlmodel==0.0.6
uvicorn==0.18.2
